[Draft] Group object for dimension styles
