+++ src/Mod/Draft/draftviewproviders/view_dimensionstyle.py
@@ -32,6 +32,21 @@ from Draft import _ViewProviderDraft
 from PySide.QtCore import QT_TRANSLATE_NOOP
 import draftutils.utils as utils
 from pivy import coin
+from draftviewproviders.view_draft_annotation import ViewProviderDraftAnnotation
+from draftviewproviders.view_draft_annotation import ViewProviderAnnotationStylesContainer
+
+
+class ViewProviderDimensionStylesContainer(ViewProviderAnnotationStylesContainer):
+    """A View Provider for the Dimension Style Container"""
+
+    def __init__(self, vobj):
+        super().__init__(vobj)
+        vobj.Proxy = self
+
+    def getIcon(self):
+
+        return ":/icons/Draft_Annotation_Style.svg"
+
 
 class ViewProviderDraftDimensionStyle(_ViewProviderDraft):
     """
